Use automake to run tests
There was custom makefile used, but integrate the test
to automake's "make check".

Create a separate "main" file for the testsuite.
The idea is to keep a standalone runner of all the suites
in the "tests" binary, but make it possible to run
separate suites with the automake machinery

This patch adds generated build/test-driver
for consistency

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>