log.c: do not move to inited state when debug() first called
If logger is in the default state, it should be inited
implicitly by the config value when fglog() is called
for the first time. If debug() is called before, it must
use stderr for debug, but leave the logger in the uninited
state.

The implementation triggered the inited flag without
actual initialization what causes the logger to stay
in the default state forever (the default log() handler
checkes the flag).

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>