rfcheader: prepare for transparent RFCHeader type
It would be better to keep parts of the message context
local instead of common static.

This is the first part of tranformation: declare
the transparent (for external code) type and take it
in use inside of rfcheader.c implementation. The context
is still kept global and static.

Since some parts of the code pass the context but use
the concrete Textlist type for that (mostly mime),
some functions are kept as is and will be refactored later.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>