debug output: make cf_get_string() produce debug
Debug output for option lookups was done when needed
every time when the lookup performed. It causes a lot of
code duplication.

This patch makes the cf_get_string() to produce debug output
itself (with parameter if exists, or without if it is blank string).

The actual output changes a bit, since it wasn't done earlier
for _every_ lookup (but for most), so now it will be for
every. In some cases the value was printed after some handling
(number conversion or adjusting for min/max). I kept the existing
debug as additional in that cases.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>