gate: change restricted address lookup logic
The HostsRestricted option was looked up once on
gate programs startup and the value was saved as
a static state inside of "address" module.

There is no real reason to use the static state,
since the check is done in one place, where the config
lookup can be done itself.

The patch gets rid of the 2 functions to access the static
storage and moves the option value to local variables,
which initialized by the config option lookup.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>