Further version fixes
The version logic is quite broken, in sense, that
the version is generated in autotools, there is also
CVS revision handling (version.c parses that explicitly),
and version.h variables.

It's better to replace version string, introduced by
commit 7f3ca21631c1f3a7a18867e71356eefa7f2e0a9b with generation
of a git hash and then use it for EXTRAVERSION. The latter
is used to generate version output in the utilities.

It usually tries to use VERSION, which was supposed to contain
CVS revision for local file version appendix generation,
but it does not fit git well. I haven't changed yet that local
version code, but it doesn't make any harm, just useless
at the moment.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>