configure.ac: fix newsbindir and yacc detection
Before news detection started, the script checked, if value of
--with-newsbindir is executable, which is definitely not for a
directory. Since later it uses innconfvar binary from the
directory, perform the check of this binary.

AC_PROG_YACC allows to continue if yacc is not detected, but it
breaks the build. So, produce AC_MSG_ERROR if it was not find,
otherwise build fails.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>