configure.ac: do not install with user/group
The INSTALL_* programs were overrided to use -u/-g during
installation, what prevented to do installation as user (into
local dir). There is check in the Makefile, if we install as
root, it will amend the owner.

The patch removes INSTALL_PROGRAM and INSTALL_SCRIPT overides,
but keeps mode (-m) for INSTALL_DATA.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>