configure.ac: fix AC_ARG_ENABLE calls
The AC_ARG_ENABLE calls are broken a bit in a way, that if you
call configure with for example --enable-option for by default
enabled options witch expected to be called with --disable, it
will be disabled since the branch called if option is given,
means not default. It require to check argument.

Add the argument check for the options and coments from the
header for them.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>