fidogate.h: adopt conditions from acconfig.h
There were some conditionally defined variables in
acconfig.h/config.h.in:

mean that the stuff under condition should be checked only if
precondition is satisfied. Later in the code the external
condition is not checked.

It's not convenient for autogenerated config.h.in, moreover
depends of the order. So moving the logic into fidogate.h.

The patch adds amendments of the detected variable accordiing to
the variabls of the external condition (there was already similar
logic for HAVE_LINK. It requires negated contition is this case)

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>