configs: copy installable configs to configs/
The example configs are installed from doc/ dir, which contains a
bunch of different examples from different people. It's not
clear, which will be installed and why. Also if I want to change
an installable config file, it's quite a shame to change somebody
else's examples, since they do not belong to me.

The patch copies config installation into the config/ directory
under the root. The previously generated files are generated as
before. There were also several configs for other programs
(sendmail, postfix, leafnode), which were generated in doc/, it
hasn't been changed.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>