autotools: remove autogenerated files
It's pretty painfull to maintain the generated files during
development, so removing them from the main branch. For the
release purposes they will be maintained in some parallel branch.

Now to build the stuff you need to run 'autoreconf' first.

The patch also adds Makefile.in to .gitignore and fixes it a bit.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>