Revert SNPRINTF/SPRINTF checks
This commit is in some sense revert of the printf part of the old
CVS commit:

2003-12-01  Andrey Slusar <rusfidogate@users.sourceforge.net>

        * src/gate/ftn2rfc.c: fix UseOriginForOrganization
            (thanks to Anton Leer)
        * package/freebsd: update
        * config.h.in: SNPRINTF change to SPRINTF
        * acconfig.h: SNPRINTF change to SPRINTF
        * src/areafix/areafix.c: SNPRINTF change to SPRINTF
        * src/common/misc.c: SNPRINTF change to SPRINTF
        * src/common/textlist.c: SNPRINTF change to SPRINTF
        * src/common/tmps.c: SNPRINTF change to SPRINTF

(after conversion to git it got id
1985ecdd8dce0ed36f5cf16a86355f4ae6683e00)

I do not know the reasons why it was done, because for me
snprintf() is preferable if it exists, but the commit makes the
code use sprintf if it exists regardless of snprintf then.

So, the patch reverts the change and uses snprintf() if it is
discovered.

Since one of the required, a check for it added to configure.ac

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>