Makefile.am: tick,toss: use relative path for external sources
The directories use $(top_srcdir) variable in the *_SOURCES lists
to refere external (not under the current directory) files. But
the filelist is processed by config.status to create directories
for dependencies and it does it relative to the current one, what
causes it to try to create src/toss/$(top_srcdir)/src/areafix/
directory (where the make variable is not expanded, it's a shell
script).

I do not know yet how to mention such files properly, so just
replace the variable with the relative path (../../src).

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>