constants.h: make file modes group writable
Since fidogate can be configured to work together with the local
news system, it may be necessary to share files in the spool. If
fidogate and news are working as different users (ftn and news
like in debian), it's natural if the common files are group
writeable.

The patch affects constants PACKET_MODE, BSY_MODE, FLO_MODE,
DIR_MODE, FILE_DIR_MODE, CONF_MODE.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>