Makefile.am: installdirs: block chown if there is no user
There is a check for root, but fakeroot makes it to pass. To make
it possible to build on a system without the user (build
package), add also a check for user existance. The actuall owner
should be then fixed externally.

Also split the command line into two, there is no big point to
keep it as one shell call, but line information from make is more
useful now.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>