constants: increase MAXMSGHEADRLEN from 70 to 71 chars
Strange logic behind the define. It supposed to limit string name
length in rfc2ftn for subject, name to and name from, and it does
it for the internal message structure. But the field
sizes (http://ftsc.org/docs/fts-0001.016) are pretty different:
36 bytes for names, 72 bytes for subject, null-terminated
strings. There is no check in the final pkt string writing, so it
should be possible to overwrite the limit for names. Also the
name handling logic looks pretty strange for me.

For the subject, because of the real string length limit is 71
characters, it is greater then the limit of 70, so it is replaced
with "(no subject)" by the checking code.

So, increase the limit to the actual subject's one.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>