toss/ngoper: redirect stdio to /dev/null instead of closing
If I understand the original idea correctly, it is not welcome to
see output of the ctlinnd in this case, because it can be
misleading. But in general doesn't look correct to close stdio
descriptors (0, 1, 2). And finally it caused problems,
ngoper/ctlinnd crashed in some special circumstances (lxc
container, selinux).

Instead of closing, let's do it correctly and redirect to
/dev/null.

Reported-by: Alex Shupikov
Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>