mime: handle mix of mime/non-mime word in subject
A problem happens with the initial implementation if email client
replies to a message with a non-7-bit subject: it inserts
typically "Re: " as normal 7 bits and then encodes as mime the
rest, but the initial implementation considers the whole subject
as plain text if it is not mimed at the beginning.

The patch adds handling of the subject word by word, where words
are space separated.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>