mime: change fix of mix string handling in non-invasive way
The previous fix changes the original line, which is not always
good and there is no real reason for that. Meanwhile it makes the
test crash since it uses RO strings as the source.

Change it in a way that it makes only copy but does not change
the original string.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>