constants: increase line length
Increase dramatically (to 32768) DEFAULT_LINE_LENGTH and
MAX_LINE_LENGTH as a temporary workaround of rfc2ftn line breaking
problem on UTF-8 messages.

1) it does not correspond the reallity: it checks the length after
base64 decoding, it's in the original charset, so for cyrillic text
it breaks the line roughly at the limit twice less;

2) it does not care about encoding (UTF-8 in my case) so easily
breaks the text in the middle of the encoded sequence.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>