ftn2rfc: always return area if found in areas
In case of ACTIVE_LOOKUP compilation ftn2rfc ignored area
information if it could not find it in news' active or could not
create it.

But it still gates the message with the default area information.

The logic does not make a lot of sense, use the area data for gating
if available.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>