mime: word based only 8bit headers encoder (base64 only)
Revisit the encoder. The old implementation is fundamentally broken:

- calculation of line length is wrong, as the result in multiline
header the first line is always overflows the limit;
- for plain headers it just does not work.

Make a new encoder which encodes headers word by word splitting with
space/tab (only base64 for now) skipping encoding of the words fully
7 bit.

Pass the whole header including the name to the encoder (it will
not encode 7 bit name), simplifies the gate code a lot.

7 bit parts must be wrapped correctly to fit the line limit. If it
contains \n, the whole part of spaces replaced with one space.

Since the words may be not 3 bytes mutliply and to combine 2 words
into one mimed they should be mimed "together", some logic with
saving and reusing unmimed reminder is implemented.

Skip "Path:" header even if it is long, otherwise inn generates dups
with local origin.

Pass the header for encoding regardless of the area flag to conform
RFC822/RFC2047 (no 8 bit headers), i.e. on the gate (ftn2rfc) side
the header is passed through the encoder if it's not fully 7 bit or
if it's 7 bit but longer then allowed.

Side modifications:

- mime_b64_encode_chunk() return the len of handled bytes (it
  always produces 4);
- make it (mime_b64_encode_chunk()) static since used only inside
  mime.c;
- mime_header_enc() accepts NUL-terminated line, no point to
  consider it as byte buffer since headers are definitely strings
  with special string handling;
- mime_header_enc() result is garanteed \n terminated. It makes code
  simple without reallocating and adding \n on the gate side, but in
  mime where the memory allocated;
- correct MIME_STRING_LIMIT to 76 instead of 74.

Future improvemnts:

- space handling;
- adding quoted-printable;
- make automatic selection of encoding based on length analisys.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>