charset: convert xlat_s to use iconv recoder
The function is used everywhere, so even if I do not like the static
state inside the module, it makes sence to keep the API at least for
the transition time.

Make the function to use the generic charset_recode_string(), which
recodes with iconv and fallbacks to the internal table recoder if it
fails.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>