mime: recognize utf-8 in plain headers, debug message fix
Allow plain headers in utf-8 regardless of the body charset (assume
utf-8 if it looks like valid sequence).

Otherwise logic has changed a bit. It used body charset always when
it exist. Now it will use it if it exists and not us-ascii,
otherwise assume the precompiled default. For 7 bit text any is ok,
if not -- it is anyway screwed up.

Fix debug message, it is not only for subject, but for any header.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>