autotools, charset.bin: build and install it
The file was declared as libexec_EXTRA but it does not make it build
and install. It would be nice to declare it as pkgdata_DATA, but it
does not solve all the problems with the file since it will be
installed to $prefix/share/fidogate and fidogate does not look for
it there by default, it uses $libexec. On the other hand automake
complains if it tries to install data in the libexec (pretty fair
complain).

Hack it with installing as a "script" to libexec at the moment.

In the future the internal recoding will be removed completely.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>