configure.ac: revisit news detection
Keep the old detection logic but make it possible to configure and
compile without inn installed or news privileges. Useful for testing
and packaging. Pass all the needed PATHs with configure
switches (capitalized names from innconfval):

--with-inndir -- main inn directory, PATHNEWS. PATHBIN, rnews and
  inews path dirived from it (adding /bin, /bin/inews, /bin/rnews);
--with-innetcdir -- inn configuration directory, PATHETC;
--with-inndbdir -- PATHDB, /var/lib/news on Debian;
--with-innspooldir -- PATHARTICLES and PATHOUTGOING derived from
  it.

Add

--disable-oldbatcher to keep defaults; it was used for inn 2.3-2.6.

Remove --with-rnews and --with-inews, leave it for autodetection
only. If the builder does not have news permissions (rnews is not
available for all to execute) the test fails anyway.

Take in use existing --without-rsf switch: make it possible to
disable send-fidogate in runinc if the owner is news;

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>