configure.ac: reuse newsbindir and provide variables report
The original assumption that inn binaries are under $PATHNEWS/bin is
wrong, PATHBIN for Fedora is /usr/libexec/news. Make it possible to
pass the bin directory with configure.

Fix setting of NEWSLIBDIR.

Provide final report with pathes, users and compiler flags (taken
from kmod) for better configure control.

It's planned to move most of the functionality variables under
runtime, so do not waste the report with them at the moment.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>