charset, ftn2rfc: implement CharsetAliasesFSC
For different historical reasons the same charset can be represented
by different names in FTN. Let's call them aliases. It is possible
to extract some of the aliases from the shipped and compiled
internal charset maps, but it's not convenient (they are not
supposed to be edited but missing some of still in use CP866
aliases) and there are plans to remove that functionality
completely.

So, add CharsetAliasesFSC option to define aliases to the canonical
charset name. Syntax is

CharsetAliasesFSC <canonical name> <space or comma separated aliases list>

config: CharsetAliasesFSC

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>