charset, rfc2ftn: add CharsetNameMap
If iconv does not support directly FSC name of a charset, make it
possible to use iconv for actual recoding, but map to FSC name for
the CHRS kludge. Add the option

CharsetNameMap <RFC/Iconv name> <FSC name>

It's supposed to have 1:1 mapping.

The patch removes old aliases lookup functions:
1) the code of reading old aliases is supposed to be removed;
2) the logic is a bit misleading, since the lookup stops at the
first match, but in the file there could be many aliases for the
same name. Split the logic between the two options -- CharsetNameMap
and CharsetAliasesFSC.

At the moment it is used only in one direction to put proper FSC
charset name if it is not supported by iconv. In ftn->rfc direction
the rfc charset is taken from the DefailtCharset directly.

config: CharsetNameMap

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>