util: add ftngenpkt
The utility is based on ftnoutpkt but does not depend of fidogate
config and allows more flexible tuning of the generated pkt/msg
including possibility to supply the raw message text. It has
different UI and allow the message to be provided from the command
line instead of stdin only which makes it easier to run via sudo.

As a disadvantage of independence of the config it requires more
paramaters to give via cmdline since it cannot read addresses or
origin from the config and has not access to the location of spool
directories too.

The original purpose was to check handling of non-standard CHRS
kludges.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>