rfc2ftn: honor 'X-FTN-Kludge: CHRS:'
If the header exists in the RFC message use the charset as the
output charset. This allows to send ftn messages in different
charsets without changing any configuration.

Other way is to have a separate header (X-FTN-Charset) since it's a
direct command to rfc->ftn converter, not the content itself. But
the CHRS kludge should be filtered out in that case anyway,
otherwise it's put inside as "misc" kludge.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>