ftninpost.c: reindent and fix do_dir()
Remove the block after 'for', reformat 'for' and add comment -- it
skips possible spaces after ' -f'. The logic of the condition is
changed a bit, but requires more refactoring. Since the buf is valid
only after str_printf(), assign it to 'p' only after it, on other
cases either point to the whole 'ftnin_sendmail' or exit the case
without executing anything (pr == FALSE).

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>