rfc2ftn: guess the news/mail mode if not forced by switches
Check the mode for every article, not only on starting.
As before if the news mode is forced with -b/-f/-n, operates in news
mode and with -t -- in mail mode.

If there are no that switches:
- if there is Newsgroups headers, operates in news mode,
- else mail mode.

In the mail mode if there are no arguments (mail addresses), assume
-t switch (get addresses from To/Cc/Bcc).

Output directory has to be set also per-message now, not once.
Areas and areasbbs must be initialized unconditionally.

This is useful for testing to have one command for both news and
mail tests.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>