areasbbs.c: replace BUF_COPY with str_copy in case of pointers
BUF_COPY() is supposed to be used with arrays only since under the
hood it gets the buffer size with sizeof(). It does not work with
"array + offset" case, sizeof returs size of the pointer then.

Do not use the wrapper but str_copy() directly. In practice it was
not a problem since the strings are shorter then pointer size (by
accident).

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>