packet: make pkt_get_string() skip string up to NUL
pkt_get_string() is used to read names and subject from the message
only. The fields must be NUL-terminated and not more than the
limits. In case if for some reason they are longer, the code reads
only amount of bytes which fits the buffer.

It can cause some trubbles for example for long subject of echomail:
- only part of the subject is read;
- the rest is read by body parser and considered as the body line;
- since it's most likely does not contain valid AREA: field anymore,
  the message is read as (broken) netmail (even if the body parser
  skips the subject's \0);

Make the function to read the whole subject regardless of its length
in the packet.

Change the API a bit:
- takes unsigned (size_t) as the buffer size;
- returns number of read bytes regardless of buffer size instead of
  error;

In the code there were no error checks, but now to make it the user
should compare the return value with the buffer size.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>