areas: areas_lookup: make logic of both w/ -H and w/o the same
I'm not sure what was intention to have check for node_eq opposite
for both branches.

For me the lookup by AREA should be like "area matches if aka is not
provided or aka matches" and then for -H build the particular entity
or return the entry as is otherwise.

The history is lost, looks like it was introduced some how with
BEST_AKA, but without good explanation (even the comment is not
fixed according to the code)

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>