ftn2rfc: do not put empty X-FTN-Flags
If there are no flags in ftn message (typical for echomail), ftn2rfc
add the an empty header which is wrong and breaks
software (newsreaders, inn when gets the article by nntp).

Do not add the headers if the value string is empty.

This is a hotfix, the code requires some more abstrations (wrapper
on adding any header, abstract headers handling).

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>