mime: do not break multibyte sequence into several mime words
According to RFC2047:

    Each 'encoded-word' MUST encode an integral number of octets. The
    'encoded-text' in each 'encoded-word' must be well-formed according
    to the encoding specified; the 'encoded-text' may not be continued in
    the next 'encoded-word'. (For example, "=?charset?Q?=?=
    =?charset?Q?AB?=" would be illegal, because the two hex digits "AB"
    must follow the "=" in the same 'encoded-word'.)

    Each 'encoded-word' MUST represent an integral number of characters.
    A multi-octet character may not be split across adjacent 'encoded-
    word's.

Make the encoder(s) aware of multibyte sequences (UTF-8 only).  On
encoding reserve some encoded line space to possible flushing of the
incomplete multibyte sequence and perform the flushing on
approaching end of line. The space is calculated as possible
maximum, so in the real life it will waste a little.

Try to make it not too much UTF-8 centric even if it's not very
useful at the moment. The maximum multibyte tail is hardcoded, but
maximum encoded length for that is calculated.

Overhead for 8bit charsets is minimal since all the values are zeros
for them.

Fixes #8.

Reported-by: Eugene Subbotin 2:5075/35@fidonet
Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>