rfc2ftn: skip names/subject translating on crossposting
If rfc message is posted to several newsgroups the messages are
created for all the echoareas separately. But actually it is not
started from scratch since the header's fields (name_to, name_from
and subject) are handled "in-place" (well, copied back after
handling) so come already translated (recoded) after the first
message is done.

Add a flag and skip the translation phase after it's done once. It
requires to initialize the 'pt' pointer for proper memory
management.

Other option would be to reinitialize the fields for every
newsgroup.

Reported-by: Eugene Subbotin 2:5075/35@fidonet
Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>