packet, rfc2ftn: extend name buffer size
Name fields in ftn message are handled in-place in a preallocated
buffer of size dictated by the standard. But it causes names
truncating for 8bit -> 8bit transcoding due to intermediate UTF-8
for headers (it cannot detect the final codepage at the time it
parses rfc headers).

It would be better to use fully dynamic buffers for it, but it
requires a lot of changes. So at the moment before the final output
own limit (introduce NAMEBUFSIZE enough to hold UTF-8 line) for the
buffers and truncate only at the last step.

The patch changes also the reading code since it depended of the
buffer size for its checks.

Reported-by: Eugene Subbotin 2:5075/35@fidonet
Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>