rfc2ftn: remove MAXMSGHEADRLEN logic
I do not understand exaclty the purpose of the code. Looks like it
checks length of _RFC_'s names/subject and if they are longer than
the limit (71), replace with shorter predefined ("SysOp", "(no
subject)"). But since buffers sizes used before for copying, in the
original code (before introducing NAMEBUFSIZE) it cannot be longer
than FTS limits (36 bytes for name for example), so the branch is
never taken.

Since sanitizing of the strings before pkt output was added, remove
that namebufsize completely. It cuts the string, does not replace
them with shorter variants (as guessed was the original intention).

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>