ftn2rfc, date: fix local tz to TZUTC affect
Local timezone is taken into consideration even when ^ATZUTC kludge
provided resulting in wrong time in the Date header. This happens
because parsedate() on parsing FTS date (without timezone info) adds
local timezone offset (for rfc date timezone presents there so
correct UTC time returned).

One of the problems in fixing the issue is that FTN date is
local (to sender, not to the gate) by definition and TZUTC kludge is
not mandatory.

Fix it by adjusting the message's date to contain correct unix time
in case of ^ATZUTC exists and leave it local otherwise. It breaks
invariant that msg.date is always in UTC, but keeps it compatible
with the old code.

Factor out timezone parsing.

Reported-by: Eugene Subbotin 2:5075/35@fidonet
Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>