rfc2ftn: make pretty name only for user part and only for netmail
Run the prettier, cvt_user_name(), which capitalizes the name and
converts spaces for user part of email, not for the real name.

To do it, move the call to rfc_parse() since after it there is no
information where the name came from.

Also change the alias logic: do the alias lookup for netmail only as
well since X-Comment-To is supposed to be already prepared and it
looks a bit strange to convert the name fetched from the
email. `postmaster` -> `Sysop` conversion for netmail only as well.

Get rid of static buffer, convert directly to msg's buffer.

TODO: Looks like there is no sense to run X-Comment-To value through
rfc_parse() too.

Reported-by: Eugene Subbotin 2:5075/35@fidonet
Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>