rfc2ftn: make pretty name only for user part and only for netmail
Run the prettier, cvt_user_name(), which capitalizes the name and
converts spaces for user part of email, not for the real name.

To do it, move the call to rfc_parse() since after it there is no
information where the name came from.

Also change the alias logic for 'To' field: do the alias lookup for
netmail only since for echos X-Comment-To is supposed to contain
receiver name/email itself. `postmaster` -> `Sysop` conversion for
netmail only as well.

Get rid of static buffers (to, from), convert directly to msg's
buffers.

Reported-by: Eugene Subbotin 2:5075/35@fidonet
Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>