ftnhatch.c, ftntick.c: enable separate tic password
Some uplinks are using separate passwords for FileFix and tic.
This commit adds the --enable-tick-password option to configure
and defines TIC_PASSWORD to change the password context in passwd
file. When this option is enabled, tic password should be added
as separate string to the passwd file in the following way:

tic    2:5075/35    password

Signed-off-by: Eugene Subbotin <evs@subbotin.org>