rfc2ftn: argify: free arrays
list_init() calls argify() which allocates space for the array it
returns (which is called here list).

It must be freed after usage.

There are several places where it is called. In rfc2ftn these are
addr_is_local_xpost() and acl_ngrp(). The latter keeps one as a
static state, so a function to free explicitly added.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>