rfcheader: get rid of static context
Refactor the functions to always accept RFCHeader context and get
rid of the static one.

The functionality will be used to make it possible to recode headers
with different defaults depending of 'areas' settings.

Make query functions NULL-friendly since in some places extra
headers lookups are not really needed and NULL will be passed.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>