rfcheader: make header_decode to produce new header
Change header_decode not to mutate existing header, but produce a
new one.

Some local helpers implemented:

header_new() - allocates RFCHeader and if passed Textlist, inits it.
header_append() - wrapper on tl_append() to add a header line.

It's supposed to be used when the original header is needed, when
cross posting to groups with different settings (charset) in
'areas'.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>