rfc2ftn: factor out snd_mail_prepare() and call for every group
Move common news/mail starting code to own function.

This is a preparation to handle cross posting to groups with
different settings (charset) in 'areas'

This patch changes the logic a bit, for crossposting it calls the
preparation code not once in the beginning of rfc message
processing, but for every newsgroup. To make it working the

ba71ee86c9ab ("rfc2ftn: skip names/subject translating on crossposting")

is reverted, the names/string are generated from the header before
evern snd_message() call.

Meanwhile it fixes a bug, introduced by ba71ee86c9ab, where on
splitting subjects did not have part indication.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>