rfc2ftn: move header decoding next to charset detection
The goal is to decode header only when all the info about charsets
is available. For news it's after area detection, for mail charsets
available, but recoding must be done earlier in snd_to_cc_bcc()
since it uses decoded field as a parameter then.

This makes it possible to use different encodings for different
areas on crossposting.

For netmail local charset info in snd_mail should be defined as
well, so call it again. Quick hack.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>