mime: make mime_body_dec() to produce new body
When the header will be new for every area, body has to be handled
with the new header as well, there will be no more static
state. Otherwise the code amends dynamic header and the amendments
are lost and it causes the body recoded again.

Actully for non-plain text the body Textlist anyway is created, so
the change is pretty tiny -- handle new list in rfc2ftn and copy it
for the plain text.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>