tests: ftn2rfc: use config 'SingleArticles', not -1 switch
The coming test has several messages in the pkt, so it's easy to
compare one rnews batch file. But the testing script calls ftn2rfc
with -1 switch causing producing a file per message.

Move the option to common config under testroot so it is possible to
override it with local test config.

test001 has a custom config already, so amend that as well.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>