rfc2ftn: recode headers for command line recipients
On recoding refactoring header recoding for netmail was moved to
snd_to_cc_bcc(), but the function is not called when the list of
recipients is passed in the command line which is the case when it's
called from the MTA. So, the headers left encoded.

Move recoding to the common place for netmail.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>