mime: get_mime_disposition: free allocated string but not tmps
TMPS is a temporary string. The module keeps allocated strings in a
list and at the end frees them all. Sort of garbage collection.

There is a mix of using TMPS and normally allocated string in the
if (disp != NULL) branch. mime_attr_value() makes copy with
strsave() inside. Make it the same way as in the previous branch.

Fixes: d3cb1fd285f9 ("mime: fix memleak on multipart decoding")
Reported-by: Eugene Subbotin 2:5075/35@fidonet
Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>