common: kludge: factor out kludge_from_tline
kludge_get() keeps last_kludge in the static storage, so probably
it's supposed to return the next kludge of the same name every
call.

But it cannot work:
- the static last_kludge is assigned every call;
- there is no way to reset the search.

Keep the old code as is, but factor the kludge name cutting code
out. It can be reused on manual iterating on kludge Textlist in the
future.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>