tests: mime: adopt hdr_enc_b64_wraps_reminder
Change the source and expected value to correspond the initial test
idea: it should check if utf8 sequence is broken between codepoints.

The problem, which mentioned in the 822c37a397ce commit message is
still there: after non-mime string mime encoding is
opened (mime_word_enc(), line 705):

    for (;;) {
        if (!state->is_mime) {
            mime_word_start(state);
        }

but then no base64 fits till the end of the line, so consumed 0 and
it's closed by mime_line_break().

Leave it as is for now since rolling back mime opening does not look
nice, but if it causes problems, can be implemented in the future.

Fixes: 822c37a397ce ("tests: mime: adopt tests to new b64 line length check")

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>