ftn2rfc: add FollowupTo header in case of carbon group
If there is CarbonNameGroup parameter in fidogate.conf any message
from FTN which contains the name is crossposted into local newsgroup
just with adding it into the Newsgroups headers field. It's better
to have FollowupTo header with the original newsgroup only so that
replies do not hit the carbon group.

Reported-by: Eugene Subbotin 2:5075/35@fidonet

Closes: #23

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>