areafix: arearix_init() accept mode as string: af, ff or tic
If compiled with --enable-tick-password there is a problem with
autocreate since areafix_init is inited with "ff" context and
password check fails since current context is "tic" one.

Make it possible to provide tic context to areafix_init().

It requires to add keys like for ff to passwd to make autocreate
working ($ key).

Get rid of MY_CONTEXT define, it duplicates global variable.

Use already existed global variable 'areafix' for mode, not
my_context string. It compared with 'ff' and now it should work for
both 'ff' and 'tic'. It may make sense to revert the condition and
use 'af' to compare and get rid of extra variable ('areafix').

Fixes: 1253fb76866d ("ftnhatch.c, ftntick.c: enable separate tic password")
Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>