msgid: dbc: save only last reference to DBC history
If 'References' contains several references rfc->ftn msgid generator
takes the last one (replying to) and calculates its CRC2. But then
it puts into db the record:

CRC32 -> All the references

So, it puts not only the last one, but the whole field, which does
not sound correct since it looks like it must be one to one
relation since the DB is used for Message-ID as well and in theory
it can fetch all the references when it does Message-ID lookup.

In practice it does not happen because CRC32 -> Message-ID record is
put first.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>